requirements.txt
Django>=4.2,<5.0
gunicorn
djangorestframework
django-cors-headers
whitenoise
Pillow
razorpay
requests
python-decouple
boto3

# RAG Chatbot dependencies
langchain
langchain-community
langchain-groq
langchain-text-splitters
langchain-huggingface
faiss-cpu
sentence-transformers
python-dotenv